file reorg, readme
