--- Sources/PinballMap/Models/LocationMachineDetails.swift
@@ -1,12 +1,0 @@
-//
-//  LocationMachines.swift
-//  pbm-swift
-//
-//  Created by Isaac Ruiz on 2/15/25.
-//
-
-import Foundation
-
-public struct LocationMachineDetails: Codable, Hashable, Sendable {
-    public let machines: [Machine]
-}
